pyproject.toml
[project]
name = "ws-server"
version = "1.0.0"
description = "Django Channels WebSocket server with LangGraph chatbot"
requires-python = ">=3.12"
dependencies = [
    "Django>=4.2,<6.0",
    "channels>=4.0,<5.0",
    "channels-redis>=4.1,<5.0",
    "daphne>=4.0,<5.0",
    "redis>=5.0,<6.0",
    "python-dotenv>=1.0,<2.0",
    "dj-database-url>=2.0,<3.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-aws>=0.1.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "asyncpg>=0.29.0",
    "langgraph-checkpoint-postgres>=0.1.0",
    "Jinja2>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
